# Improve `WindowTransform::updateAggregationState` by `addBatchSinglePlace`