@novikd 2023-02-15
Please add performance tests: https://clickhouse.com/docs/en/development/tests#performance-tests