@lgbo-ustc 2023-02-21
There is a failure in `test_storage_rabbitmq/test.py::test_rabbitmq_drop_mv`, seems not related to this pr.