@alexey-milovidov 2023-12-30
Merged in https://github.com/ClickHouse/ClickHouse/pull/56120.